services/auth/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
redis>=5.0.0
httpx>=0.26.0
python-json-logger>=2.0.7
